Trying to get AHI player working, trying to track down error, and a bit translation.
More of the comments are translated to English.
Trying override the playergroup, by inserting the player into the source
code, as suggested,
(I'm getting an error says "trouble with player group" if AHI is enabled
in hippo prefs, on OS4, I'm trying to track that error down.)
I see no point of smoke and mirror stuff, I was thinking it be nice
clean up a bit delete encryption code, for now I just made it compiler
option you enable and disable.